move bplan component to its app
