+++ meinberlin/apps/bplan/views.py
@@ -4,6 +4,9 @@ from django.views.generic import TemplateView
 
 from adhocracy4.rules import mixins as rules_mixins
 from meinberlin.apps.contrib.views import ProjectContextDispatcher
+from meinberlin.apps.dashboard2.components.forms.views import \
+    ProjectComponentFormView
+from meinberlin.apps.extprojects.views import ExternalProjectCreateView
 
 from . import forms
 from . import models
@@ -28,3 +31,30 @@ class BplanStatementFormView(ProjectContextDispatcher,
 
 class BplanStatemenSentView(TemplateView):
     template_name = 'meinberlin_bplan/statement_sent.html'
+
+
+class BplanProjectCreateView(ExternalProjectCreateView):
+
+    model = models.Bplan
+    slug_url_kwarg = 'project_slug'
+    blueprint_key = 'bplan'
+    form_class = forms.BplanProjectCreateForm
+    template_name = 'meinberlin_dashboard/external_project_create_form.html'
+
+
+class BplanProjectUpdateView(ProjectComponentFormView):
+
+    model = models.Bplan
+
+    def get_project(self, *args, **kwargs):
+        project = super().get_project(*args, **kwargs)
+        return project.externalproject.bplan
+
+    def get_object(self, queryset=None):
+        return self.project
+
+    def validate_object_project(self):
+        return True
+
+    def validate_object_module(self):
+        return True
